Fuzzy_Multiplex
Few words about the model and the following script; this model is a bidirectional co-evolutionary soft targeting hierarchical fuzzy multiplex optimizer. Following are three distinct cases that illustrated how the same concept can be utilized for different occasions in a simple and elegant manner and a few words about the model.

The optimal state of the model is achieved when it converges to the objectives, yet does not overfit inner data patterns. To achieve that soft argets are derived from the outer layer and are in place constraining the inner layer activations to optimal patterns, but not optimal fit, i.e overfit. Addionally, the model optimizes objectives simultaneously-dynamically in the inner layer and then then at the outer, based on external criteria. Lastly, during the optimization process the inner layer structures act indipendendly and evolve dynamically with the outer nodes, yet another fuzzy or logical constrain is imposed; the evolution of each inner FCM is evaluated against its peers on a local level and is penalized or rewarded based on its information merit.

Finally, the model creates multiple good instances of the targeted data; at the processes-outer layer level an activation, i.e. the best approximation, is selected. Furthermore, other top inner layer imitations are used to define a fuzzy interval IVFS (internal valued fuzzy set) for each outer node instance corresponding to each inner activation; that can be acheived at an arbitrary level and upon implementation it creates soft-target activations.